(chore): refactors cluster controller utils
